Remove MsbmsSystAuthentication instance test data
When first created the test data included a copy of the test data from
MsbmsSystInstanceMgr since we depend on that for some references.

Rather than copy that data into the tests here, simply reference the
unit test from the MsbmsSystInstanceMgr in the MsbmsSystAuthentication
unit test subplan directly.

The positive side of this is if we re-work the Instance Manager, we'll
immediately get the re-worked test data in the Authentication module
tests.  The downside of course is that this creates a non-obvious
dependency between testing data sets.

I think the balance of the issue lies in favor of not duplicating data
right now.  Even in the independent test data approach, a change to
MsbmsSystInstanceMgr would mean hunting for where all the test data
might live.  In the shared data approach, a breaking change made to the
test data will at least cause dependents to fail their tests which would
would give notice as to where the sensitivities to the change exist.

So not clear this change is correct, but reasonable enough for now.

This change therefore:

  - Removes the copied MsbmsSystInstanceMgr data from the
    MsbmsSystAuthentication test data.

  - Adds a reference to the MsbmsSystInstanceMgr test data from the
    MsbmsSystAuthentication unit testing buildplan.

  - Incrementally adds some new testing data to MsbmsSystAuthentication
    though not directly related to the primary purpose of this commit.