Refactor database directory structure
This refactoring makes the database structure resemble the application
code structure more closely, as well as removing the prefixing of
module kind (e.g. app_ = application, cmp_ = component) from the name
itself.  This reduces naming conventions to a single convention
(almost).

The only place this older naming convention is kept is in naming the
buildplan files.   This is because they want to live in the root
data directory and knowing how the specific buildplans file relates to
the other more discrete directory structure is helpful.  This may be
rethought; really how badly does the buildplans file need to be in the
root?  But for now this is better than what we had and even what we had
was workable.