Update Elixir documentation handling
This moves the default location of the ExDoc generated documentation for
all Elixir code to documentation/technical/app_server.  There's still
work to do, such as getting cross component documentation references to
not try to get to hexdocs.pm, but it's a good start.

Also add a build_docs script which will build all the documentation.
Eventually this will include database documentation as well.