Refactor MsbmsSystInstanceMgr test privileges
This is the database privilege simplification where we'll rely on the
privileges defined in dependency tests to set privileges for this
module's tests.  That should avoid some duplication and be more
resilient to code changes in the dependencies.

MsbmsSystInstanceMgr is the only module other than
MsbmsSystAuthentication (handled elsewhere) that currently can be
simplified in this way.