Fix bad syst_identities trigger validation query
When adding identities the validation would allow two identities for the
same account_identifier for the same access_account.  That shouldn't
happen.

The issue was were were checking for pre-existing duplicates in all
access accounts in the resolution scope except the target
access_account.  This needs to be all access accounts in the resolution
scope.

Also add note for PostgreSQL 15 updates to unique index handling.  That
may well eliminate the need for this trigger altogether.