Fix syst_identities unique trigger timing
This trigger was designated as a BEFORE INSERT AND UPDATE trigger, but
the UPDATE case it prevented valid updates to the record.

The fields upon which the uniqueness criteria is based are designated as
immutable (see msbms_syst.trig_i_u_syst_identities() trigger).  This
means they can't be changed to a non-unique set of values so long as
they are inserted correctly.

So just check the constraint on INSERT and not UPDATE.  Yes, someone
could do a direct insert into the database and arguably we should handle
the update case, too.  But it's unlikely and the app won't (be able to)
do it.