Refactor SystCredentials/SystIdentities column
The external_name column defined on syst_credentials was intended to
give certain credentials, such as API tokens, a name so that a user
could manage credentials such as application specific credentials.

This was a fine idea, but it doesn't belong on syst_credentials, it
should go on syst_identities.  We have to have the identity records
since we won't be able to know what the credentials actually are; we
should remember that syst_identities is the table where we say who we
are and syst_credentials are how we prove it.  Without the "who" part
we'd be trying each password in the database to see if we could match
the hash and that would be too resource intensive and too slow.

So, these API token class credentials will be tied to a specific
identity and the identity record data isn't to be considered secret; it
could well be sensitive, but there are no guarantees that that data is
save from attackers.  So this is where our external naming field goes
with the same assumptions.

(note: we could encrypt the external_name field so that a user doesn't
give away anything too juicy, but for now this will suffice.)