Add MsbmsSystAuthentication DevSupport
For interactive development, it's beneficial to use the testing related
database setup to build/start the database and services so that
iex -S mix can be used.

Create a near copy of the TestingSupport and TestHelper functions which
create the database based on testing seed data and start the services,
including the Enum services, so that full interactive development is
a possibility.

This is a "good enough" implementation of the ability.  The next steps
would be to create a module which can just be included into dev builds
as a dependency, but which can use a configuration file to for the
one or two items which are specific to the component being developed.
Finally, de-dupe the code so that the testing support used the same
functionality to bootstrap the test run environments.  All later...