Initial commit of base license
Right now the intent is to initially license the software on a personal,
non-commercial use license.  This should be sufficient to allow others
to examine the software without compromising the options we might want
to pursue later.

To this end, start with the PolyForm "Strict" license.  This is still
more permissive than we want, but is a more than workable starting point
given the budget for license development.