Remove credential_id from AccessAccount/Instance
The syst_access_account_instance_assocs table includes the idea to very
granularly permit only certain credential methods to be usable with a
given instance.  Remove that.

The reason is that first, the idea isn't quite right: if anything we
should be more interested in authorizing certain identification methods
which implies credentialing methods (e.g. email -> password).  Also, the
access account/instance associations table probably isn't the correct
place for this either.  If we really want to do this then we'd want an
owner level acceptable identification methods table and a separate
owner level acceptable credentialing methods table.  This would reduce
duplicate data and better express the likely needs for such a feature.

For now, we'll avoid the complexity and just eliminate the association
and leave it open as to how we might handle the fine grained expression
of approved authentication mechanisms.