Refactor MsbmsSystDatastore.DbTypes; add compare
For MsbmsSystAuthentication, we're going to need to evaluate different
things like global and owner password length rules and do so in
comparison with each other.  Currently there is no way to do this other
than manually each time or by delegating the comparison to the database.

This commit refactors and reorganizes the DbTypes area to support
comparisons of range types and range types with related non-range types.

Note that the implementation seems complex/naive for what's actually
happening.  Also, the implementation chains function calls together in
the function body rather than allowing data to be passed from one
function to another (i.e. via pipe operator).  In the end I don't think
it will matter since the opportunity to compose these functions is
small, but it still just feels expedient rather than correct.