Refactor DbTypes.TimestampRange to DateTimeRange
Originally this was named to be consistent with the database naming of
the type since these were DbTypes and more directly informative about
what is being done.  However, in practice, it's more annoying and
confusing when actually working with the type in Elixir since the
un-ranged type is DateTime.

This commit makes the naming internally consistent with the Elixir 
naming.  In retrospect this seems obvious: we're building something 
we're going to be working with in Elixir so we should be consistent with
Elixir naming.