Refactor DbTypes protocol to DbTypes.Range
It didn't feel right having range specific protocol definition at the
root of DbTypes.  It's conceivable that we'll have non-range custom
types that could have their own concerns that wouldn't apply to Ranges.

So this commit makes the protocol definition more specific by making the
protocol definition DbTypes.Range.  This file is included with the type
definitions which I don't like (two different kinds of things grouped
together), but the API feels right and sitting here now it feels more
appropriate to following the file path to module name convention at the
cost of better organization in the file system.