Add DbTypes.Range protocol
This brings back the DbTypes.Range protocol except this time it is
addition to the DbTypes protocol and actually covers functions which
have no meaning outside of range comparisons.

Specifically it became immediately apparent that the coarse grained
returns of the compare/2 and test_compare/3 were insufficient in some
cases to make actionable decisions.  For example, testing owner password
length rules against the global baseline could result a `:lcr`/`:rcl`
value when the owner rules were both sufficient and insufficient.  In
these cases we need to know the details of the lower and upper
comparisons and this is what the new DbTypes.Range protocol deals with.
It adds the `bounds_compare` function which returns a map of the lower
and upper bounds comparison operators for those more complex needs.

Implement for all supported types; DbTypes.Interval is not a range type
so it doesn't get this protocol.  Also implement tests for the new
functionality.