Add upper/1 & lower/1 to DbTypes.Range protocol
Add implementations as well.  One weakness right now is that the
bounds_compare/2 function of the protocol applies nicely to the
supported base types (e.g. Integer, Date) as well so we can work with
mixes of base and range types, but upper and lower make less sense.  BUT
we need to implement upper and lower for the base types since they
implement the Range protocol for that useful case.  Now upper and lower
just return the base type value, but really it's the first sign that the
protocol design isn't quite right.