Add syst_credentials trigger to enforce uniqueness
Create a trigger to enforce the required uniqueness conditions of the
credential table.  This must be a trigger rather than a unique
constraint because the credential_for_identity_id must treat NULL as
being equal to all other NULLs and PostgreSQL 14 and earlier don't
treat NULLs as such in unique constraints (all NULLs are individually
unique).

When PostgreSQL 15 is released it is set to include a feature for
optionally treating NULLs as non-unique in which case we can replace
this trigger with a simple unique constraint.