Refactor MsbmsSystInstanceMgr typespecs
This was done while trying to solve Dialyzer complaints.  In the vast
majority of typespecs we don't name parameters but rather then specify
the expected types.  In MsbmsSystInstanceMgr calls there were some
disparities with this standard which are fixed here.

To be fair there's a good argument that we should have refactored the
other direction, adding names to all the unnamed parameters in
typespecs; this is especially true since we're really adding typespecs
for documentation purposes.  Still, I think the costs outweigh the
benefits.  We get more work, more duplication, and even noiser code
since the typespecs would be longer.  Given that most tooling and
documentation generation puts the typespec in close proximity to the
actual function signature which uses the parameter names, we wouldn't
get a terribly big win for the extra effort.

The only places where we'd come out ahead are in more complex, multi-
parameter functions calls.  In some of those cases, both the typespec
and the function display with its parameters make it difficult to relate
the displayed typespec and displayed function call by eye.  But those
kinds of functions are fairly rare.