Add MsbmsSystAuthentication.Impl.Hash module
This module is a wrapper around other hashing algorithm libraries.  By
wrapping the other libraries we can change those pieces out without
necessarily disrupting where credential processing functions need to
make use of the functions.  For example, Argon2 can have problems in
building on Windows so we might make the credential hashing algorithm
used dependent on build environment; in those cases we'd have different
(though similar) calls from different libraries.

The thing we need to be careful of with this approach is that we need to
be clear of purpose in naming our exposed functions.  So credential
handling functions should be clear that they're for use in credential
handling and those functions not suitable for use with credentials
should be named so that they would raise flags when used in credential
processing contexts (i.e. function `weak_hash/1` is such as example).