Add syst_credentials trig, delete linked identity
When a credential is linked to a specific identity via the
syst_credentials.credential_for_identity_id, the relationship between
the identity and credential is very tight and there isn't currently a
case where such an identity type should exist without a corresponding
credential.

When deleting the identity, the foreign key relationship defined on the
syst_credentials table takes care of deleting linked credential via
ON CASCADE DELETE but there is no corresponding automatic deletion of
the identity when the credential record is deleted first.  This trigger
solves that problem allowing you to delete either while having a
consistent outcome.

This is an after trigger to avoid having extraneous processing due to
the cascade of the foreign key relationship; the credential record
should already be gone by the time this trigger fires.