Add identification predicate :allow_unvalidated
The query building for looking up specific Access Account by identifier
previously could build queries requiring validated identities or
requiring unvalidated identities, but not both.

In the user authentication extended logic, it can be useful to tell a
user that their authentication is being rejected because the account is
not yet validated.  To support that case we still need to use the
identity to find the access account, but we'll decide the authentication
status to set rather than simply reject it outright.