Fix MsbmsSystAuthentication date related test data
This is specifically related to the changed name & meaning of the now
named `identity_expires` column.  Previously the meaning would have us
expect a value in the past which wouldn't hinder operations so long as
the record was validated.

Now, however, the meaning is whether or not the identity itself is time
limited and a date in the past in `identity_expires` means that the
identity itself is expired and no longer usable.  Since this test data
is suppose to be usable it is by definition "not expired".