Add password history testing data
In MsbmsSystAuthentication, support for prohibition of recently used
passwords exists and needs to be testable.

This commit adds some basic testing data related to testing that
functionality.  It's unclear whether this will be sufficient or final
(its not expected to be either), but for now it's at least a
placeholder.