Add MsbmsSystAuthentication ExtendedAuthLogic mod
This module implements high level, composite logic drawing on the more
dedicated functional areas such as identity and credentials.  The
central idea being that the higher level interface to the module should
abstract away some of the distinction between the various authentication
related areas of identity, credential, password & network rules, etc.
As such the broad processes of authentication should be implemented in
this module.

Also, a fairly substantial update and extension of the Types module is
also included in this commit.  These updates and additions are all
related to the extended authentication logic module implementation.