Refactor/Add Identity expiration management funcs
In MsbmsSystAuthentication.Impl.Identity, there are generic functions
which deal with the syst_identities.identity_expired column data.  The
function set was incomplete in that you could set/change an expiration,
but not clear it altogether which is a valid operation.  To correct and
reduce duplication of effort the following changes are made:

  - Add a clear_identity_expiration/1 function to clear the expiration
    column when needed.

  - Refactor the set_identity_expiration/2 function, removing some of
    the basic lookup and updating functionality.  Replace the lookup
    functionality with a call to a new generic record retrieval private
    function.  Replace the full update call with a simpler call to the
    Helpers.update_record/2 function which serves the same purpose.

  -