Fix Impl.Identities.identity_expired?/1 nil bug
This function compared current date/time with the
syst_identities.identity_expires column value.  This works fine unless
the identity_expires value is NULL.  This bug may originate with the
original name/meaning of this column; I think that value was always set
so the NULL value problem wouldn't arise.

Add a check to return false if the value is NULL/nil since that means
the identity never expires.