Refactor MsbmsSystUtils.Impl.resolve_options/2
This function used the `||` operator to determine if the caller provided
a value for the key or if the default value should be used.  The problem
with this operator is that it makes its determination of the truthiness
of the first argument; this means that a user option that is a boolean
`false` will always prompt the default value which, if boolean `true`
will be wrong.

Write the function to be more explicitly dependent on the nil value
being the only case where the default is used.  Also add clarifying
documentation to the API function which fronts resolve_options/2.