Reorder syst_identities table definition elements
Originally the `syst_identities_primary_validator_chk` check constraint
also tested what was the identity_expires column.  This is no longer the
case so the ordering of the columns and the constraint definition are no
longer really correct.

Reorder the constraint location so that it immediately follows the
columns which it governs.