Refactor Impl.AccountCode.reset_identity/2
This function originally tried to update the
syst_identities.account_identifier column directly with a new
AccountCode, but the database does not allow that column to be updated.

Correct the process here so that a reset is a delete of the old
syst_identities record matched with the creation of a new record.