Remove Identity.Recovery validated_identity_id
In the record creating function, the validates_identity_id record value
was being set but this is not appropriate.  Recovery does look up an
identity (i.e. email address) to locate the access account which needs
the password recovery, but the email identity itself doesn't need
recovery.  This is likely a copy/paste error from Identity.Validation
where the email identity is actually the target of validation.

Remove the reference to validated_identity_id