Fix bad Identity.Recovery params in function call
In Recovery.identity_access_account/2, the call to
Helpers.get_identification_query/4 passed the :require_validation value
in essence asking that the recovery identity itself was validated.  The
system doesn't permit the recovery identity to be validated so the
resulting final query could never succeed.

I actually tried refactoring both Recovery & Validation to count their
identities as being validated by default... but the database rules
disallow this.  Clearly I've waffled on whether or not validation should
only be recorded on only records that can be validated (like email
identities validated by validation records) or if any record that can be
processed should appear as validated.

For now we'll stop before refactoring the database and just run with the
thinking in effect at the time when the rule was put into place: that an
identity record like a Recovery record cannot be validated since the
user doesn't validate the Recovery record itself (same for Validation
Identity records).