Fix Credential.GenericToken bad params handling
When built, Impl.Credential.GenericToken.get_credential_record/3 was
intended to either return the record, or return nil if the parameters
were bad or incomplete.  This happens, but causes non-fatal errors to
appear when one of the parameters is nil since Ecto wants an an explicit
nil check in the where clause of the query rather than an equality
check whereas the technique tried here would be perfectly acceptable in,
say, an PL/pgSQL function.

To avoid Ecto complaining about this, don't try the query at all if one
of the parameters is nil.  This is probably slightly more efficient
anyway.