Fix Credential.GenericToken rebinding issue
The delete_credential/2 credential_type parameter is needed in both its
atom form for Elixir handling, but also string form for the Ecto query.
Early on we take the credential_type parameter which comes in the atom
form and rebind credential_type as a string.  This worked OK to get us
through the query, but was a source a of subtle failure later because
we'd return the string form when we really needed the atom form again.

While rebinding can reduce the number of names you need to mentally
track while working with the code, rebinding across types like this
feels error prone and probably should be avoided.

The correct answer for this situation is just create the extra variable
at the time of casting; In Erlang we would have had to do this anyway
and the alternative (multiple conversion) feels like throwing fuel on
the fire.