Add TODO to test_data.eex.sql for better coding
Even though this TODO is only being added to MsbmsSystAuthentication
test_data.eex.sql, it really applies to more than here.  The issue is
that the testing data up to this point hasn't been that bad to just
brute force into the file and the database.  But as the functionality
gets more complex and with increasing numbers of behavioral variants in
any one function, the data to support testing will get more complex and
the brute force manually created inserts less and less comprehensible.

Really we should be describing our test data and allowing procedures to
create the data based on the description.  This way the manipulation of
test case data is focused on the description of the data needed without
the distraction of the mechanics of creating that in the database being
there.

We're almost done with this test data and the value to going back and
restructuring now isn't that great, but any significant work with this
test data should budget time to do this clean-up and payoff this debt.