Refactor MsbmsSystAuthentication test_data.eex.sql
This commit implements the to-do added in the last commit.  I should by
now just acknowledge that if something is cumbersome enough to warrant
commenting on fixing it, I should just fix it.

In this case I was going to add the next series of test data and
realizing how scattered the definitions were and how ad hoc the creation
was, it was just going to be painful and error prone.

While it will still be painful, it will be much less so and the data
itself will be clearer... so while the data definitions are still very
nit-picky, all the data is in one place and much more easily viewable.