Simply Authentication test_data validation ident
Ran into an issue where we'd get errors if any part of a validation
identity was defined without also having a completed validation
credential also defined.  Under normal circumstances having a validation
identity without a corresponding credential shouldn't happen in
practice, but can be useful in the test data since we are testing lower
level functions in which incomplete data can be the correct starting
point.

So this commit simplifies the code a bit by eliminating the conditional
logic which gated on any validation data existing or not and making the
inserts directly responsible for checking their individual conditions.
Not only should this fix our bug and make the code a bit simpler... we
should also now be a bit more declarative.