Remove Validation cred creation from Elixir test
If the test data is created correctly, there is no need of creating new
validation credential records in the delete credential tests.
Previously, with the ad hoc test data creation, find the right place and
time to create such records was more cumbersome than just creating them
using the Elixir functions, but that problem is there anymore.

Update the validation test data so that it creates the needed validation
credential records for the two delete record tests and remove the code
that created them in the Elixir test scripts.