Refactor syst_credentials unique check
The uniqueness constraint for this table depended on being able to treat
NULL values in the credential_for_identity_id column being considered as
not unique to other NULLs in that column.  In versions of PostgreSQL
prior to 15 this was not possible and enforcing such a constraint
required a trigger.

With the release of PostgreSQL 15 unique constraints can now be defined
as UNIQUE NULLS NOT DISTINCT.  This allows us to get rid of our trigger
function and add a simpler constraint with this new directive.