Refactor Password.confirm_credential/3 nil result
When confirming a credential, we typically don't want to error on
something like missing data since that can be used to disclose the
existence of an account without a successful authentication.  Rather we
want to return one of our defined bad credential codes.  The original
version of this code however would error if no credential record was
found because there was no explicit handling of the no credential case
and we had an explicit map access which failed when the variable the map
would be in was itself nil.

Change this so that we only try to get credential extended status if we
actually have a successful credential verification.  Otherwise, just
return an empty list for the extended status.