Refactor trig_i_d_syst_disallowed_passwords
Previously we were going to not provide a way via the app to remove
disallowed passwords, but thinking it over that's probably a little
draconian.  We need to protect the ability to remove disallowed
credentials, but if you can add them, you can probably add them by
mistake which means you need a normal way to correct the issue.