Refactor trig_i_i_syst_disallowed_passwords
This loosens the insert process a bit by ignoring any primary key
collisions.  Since the hashed credential is the primary key, and getting
a listing of the hashed credentials is non-trivial, it's better to just
ignore any attempt at a duplicate than to error; the user won't have
much information or help in avoiding duplicates ignoring the duplicate
seems reasonable.