Fix tests related to disallow_compromised pwd rule
Earlier a bug was fixed with boolean value defaulting and the setting of
password rules such as disallow_compromised.  The test assumed that the
buggy behavior was actually the correct behavior and that we were
overriding 'weaker' rules at create time rather than relying on read
time evaluation.  This was incorrect.

We want the write time password rules to reflect the user intent even if
the user's intention is the have a weaker password rule.  This is
because the global password rule that the owner rules are compared
against can themselves change over time; our rule is to always honor the
explicit expression of user intent if we can and so the only time we
can evaluate the owner password rules for relative weakness/strength is
at read time.

This commit changes the tests to ensure that the written value of the
owner rules reflect the user desired value and that in determining the
effective rule for a login user the stronger value between global and
owner rules is used.