Refactor NetworkRules.host_disallowed/1 return
This also includes renaming the function from host_disallowed? to just
host_disallowed.

The reason for this change is consistency, though to be frank I'm not so
enthusiastic on this one.  The key issue is that the simplicity of a
simple true/false boolean return is being lost for a more complex result
tuple to wrap the boolean; it's just a little less elegant.  Having said
all that, our library-like nature and the determination that we're
returning result tuples, especially for database wrapping functions
which can error for a variety of operational reasons means that the
result tuple is called for based on our standards (for now).  I also
worried about consistency in this regard across the app so far and I
don't want to consciously make the consistency story worse.