Refactor create_disallowed_host/1 for robustness
When attempting to add a host to the Disallowed Hosts list which is
already on the list, the database will ignore the insert attempt
(ON CONFLICT/DO NOTHING), but this returns an empty struct rather than
the record which conflicted.  That's OK, but rather than an empty
struct we should just return `nil`.