Add MsbmsSystAuthentication IntegrationTest
This starts the integration tests for the MsbmsSystAuthentication
module.  These tests will roughly follow the system build up process to
prove that a succession of steps resembling a business process can be
supported by the functionality and public API of the system.

The initial tests are related to Disallowed Passwords and Owner Password
Rules and are included in this commit.