Refactor PasswordRules.verify_password_rules/2
This is probably more than just a refactoring.  One issue that I've not
been comfortable with is when should result tuples be used relative to
just expecting a result or exception to be raise.  Given that our
modules are library-like, whether to crash or not is could be seen as
the responsibility of the caller.  Insofar as we're only library-like
and not actually writing libraries there's also an argument that much of
this ceremony should be avoided and just allowing the components to
crash is fine.

Finally, there's where I've landed for now: exposing both a non-crashing
version of the function:

  function() :: {:ok, result} | {:error, reason}

and a crashing version:

  function!() :: result

This lets the caller opt into the behavior they want, allowing them to
avoid the necessary deconstruction of the result tuple if they're just
going to crash anyway; and so far, many of the callers are just matching
on the success tuple allowing anything else to blow up.

This should allow more functions to participate in pipe operator chains
and allow simpler binding expressions in many cases at the cost of the
extra `!` character in the caller code and the small amount of extra
code to deal with wrapping the crashing version of the code in the
result tuple version.

Some other standards established here that have not been the way when
this approach was implemented previously:

  - In the crashing version of functions, still include a rescue block
    in order to log the errors.  Re-raise once the exception is logged.
    This wasn't done previously when the non-crash/crash approach was
    tried earlier, but putting the log closer to were the exception is
    happening ensures we log the issue consistently and without regard
    to call method.  This means that the rescue block in the non-
    crashing wrappers should just be `error -> {:error, error}`.

  - If the parameters are narrowly defined, include a catch-all function
    head to log and raise due to bad parameters.

This is currently the only place where the standard exists, so 
refactoring other API related implementations needs to be a TODO.  I
feel sufficiently comfortable with this approach that I'll likely 
schedule at the end of the MsbmSystAuthentication cycle.