Refactor Authentication for Result Tuple/Raise
This is a fairly broad based refactoring to allow many publicly
available functions to return either a result tuple or to simply succeed
or raise.

Our components are treated as though they are third party libraries in
many ways and in this capacity its good to offer the caller the ability
to handle certain errors rather than just crashing.  On the other hand
writing a lot of `{:ok, _somthing} =` when we'll just crash anyway is
also a little tedious.  So let's go ahead and write these public
functions to just blow up and then give a convenience function that can
wrap the primary in a result tuple.  This won't be a perfect realization
and we're not back-porting this everywhere right now, but will do so on
a case by case basis.

A summary of changes here:

 - Update the Credential protocol to require both the result tuple and
   raise of error versions of the functions where appropriate.

 - Update the implementations of the Credential protocol to respect the
   new requirements, and addressing any extra non-protocol functions
   along the way.

 - Update PasswordRules with the same kinds of changes for result tuple/
   raise on error.

 - Update related tests scripts and test data to include testing of both
   the result tuple and raise on error versions.  Note that this is very
   ad hoc and very happy path oriented.  An area for later improvement.

 - Add additional Owner records to the MsbmsSystInstanceMgr
   test_data_unit_test.eex.sql to more easily support testing in other
   modules (such as MsbmsSystAuthentication), even though these same
   Owner records won't be so helpful in MsbmsSystInstanceMgr testing...
   it shouldn't hurt.