Fix bad columns in Network Rule API view triggers
The INSTEAD OF triggers for the network rule API views used the simple
RETURNING * INTO new/old formula that usually works.  In this case they
don't work because the select includes a convenience column (ip_family)
which isn't part of the underlying tables.

Explicitly state the columns and execute the ip_family expression in all
INSTEAD OF triggers associated with the network rules tables.