Add DbTypes.Inet conversion to/from net_address
The net_address library (module `IP`) adds network addresses that look
very much like standard Erlang IP addresses, though this begins to
differ once you get into representing subnets.

This commit adds conversion to and from the DbType.Inet which the
database changesets are expecting.  It's difficult to know right now
which style will be easier to work with in client code, so the idea
right now is that it shouldn't matter much which is chosen.

Also committed here is a fix for the DbTypes.Inet to Postgrex.INET
conversion functions.  I didn't read the typespecs closely enough and 
assumed a more Erlangy representation for the netmask when, in fact, the
desired representation was just an integer count of the bits in the 
mask.