Remove syst_global_network_rules.template_rule
Originally, the functionality envisioned that owners would start with
a default set of Network Rules based on globally defined template rules.

In retrospect, I think this isn't really necessary.  The existence of
Global Network Rules, makes the desirability of defaulting a standard
set of Owner rules dubious.  It's true, we may want a set of
"recommended" rules which can be opt-out for Owners, but the extra
complexity at this stage probably doesn't warrant having this in the
table structure or recognized by the code.  So let's shelve this for
now.

Remove from the tables, the API views and INSTEAD OF triggers, and from
the Elixir application code.