Refactor Network Rule params functional_type
The functional_type attribute of the map used to create or update the
various Network Rule record types defined the functional_type as being
of type "String.t()".  This is workable, but sometimes working with
atoms is more convenient.

Rather than force a specific representation, just allow both a string
representation or an atom representation.  This should alleviate some of
the tension around which it should be.  It won't eliminate this tension
completely as the data structs will still return strings and the applied
rule function will still return atoms.

This will probably be a case of feeling out which feels more useful in
practice and then reducing the functionality to the useful case.