Fix syst_global_network_rules constraint deferral
The uniqueness constraint on the ordering column was causing errors when
inserting rules with the same ordering value.  In practice, this should
result in a re-ordering of existing records which collide rather than
actually resulting in constraint violation.

The issue was the deferrability of the column hadn't been addressed.
This commit resolves that issue.